requirements.txt
Flask==2.0.1
Flask-MySQLdb==0.2.0
Jinja2==3.0.1
python-dotenv==0.19.0
